Add plugin support to run NDSes for data files
Add a new mechanism to construct command lines to run an NDS program
with a selected data file as the argument. Plugins are placed in the
`/nds/plugins/` directory with the file extension they handle as the
name and an nds extension. E.g. copy `tuna-vids.nds` to
`/nds/plugins/avi.nds` and then you will be able to browse for and play
`.avi` files.

Also refactor main() to move `.argv` support to its own function.